+++ Riot/Modules/Contacts/Details/ContactDetailsViewController.m
@@ -727,7 +727,7 @@
         
         if (indexPath.row < actionsArray.count)
         {
-            NSNumber *actionNumber = [actionsArray objectAtIndex:indexPath.row];
+            NSNumber *actionNumber = actionsArray[indexPath.row];
             
             NSString *title = [self actionButtonTitle:actionNumber.unsignedIntegerValue];