+++ Riot/Modules/StartChat/StartChatViewController.m
@@ -257,7 +257,7 @@
         if (identifiers.count)
         {
             // Here the contact can only have one identifier
-            [contactsDataSource.ignoredContactsByMatrixId setObject:contact forKey:identifiers.firstObject];
+            contactsDataSource.ignoredContactsByMatrixId[identifiers.firstObject] = contact;
         }
         else
         {
@@ -266,7 +266,7 @@
             {
                 // Here the contact can only have one email
                 MXKEmail *email = emails.firstObject;
-                [contactsDataSource.ignoredContactsByEmail setObject:contact forKey:email.emailAddress];
+                contactsDataSource.ignoredContactsByEmail[email.emailAddress] = contact;
             }
         }
         isMultiUseNameByDisplayName[contact.displayName] = (isMultiUseNameByDisplayName[contact.displayName] ? @(YES) : @(NO));
@@ -274,7 +274,7 @@
     
     if (userContact)
     {
-        [contactsDataSource.ignoredContactsByMatrixId setObject:userContact forKey:self.mainSession.myUser.userId];
+        contactsDataSource.ignoredContactsByMatrixId[self.mainSession.myUser.userId] = userContact;
     }
 }